added create and monitor
Added a create and a monitoring function to the views.py. This functions arten't tested yet but the implementation/setup of this is needed for the urls.py Mr. Heinemann is working at.

Another change is the cleanup of the settings.py on main package. I've deleted some not necessary comments and also some settings we won't need.