Change convoy to reference for addresses
Renamed the path convoy and gave this service a better name: 'address'

The functions are set up but have no direct impact yet. This service will return the current state of the addresses if you give it a 'truckId'. The URL-resolving has also been updated, as well as the settings.py, etc. because the renaming needed an update of this files.

PS:
I'll implement this functionality this eve, so you might test it soon