Query and Initialization finished
The Querys are complete, so every truck can access the needed data from the truck in front / behind.

The initialization is set by a really curious URI. We only need to define the individual initialization (as said in the comment of the class initialization.py).

Next step will be the movement-connection.